merge

+++ src/CeesVanEgmond/Minify/Providers/BaseProvider.php
@@ -215,10 +215,15 @@ abstract class BaseProvider implements Countable
     protected function removeOldFiles()
     {
         $pattern = $this->outputDir . $this->getHashedFilename() . '*';
-        foreach (glob($pattern) as $file)
+        $find = glob($pattern);
+        
+        if( is_array($find) && count($find) )
         {
-            if ( ! unlink($file) ) {
-                throw new CannotRemoveFileException("File '{$file}' cannot be removed");
+            foreach ($find as $file)
+            {
+                if ( ! unlink($file) ) {
+                    throw new CannotRemoveFileException("File '{$file}' cannot be removed");
+                }
             }
         }
     }
